+++ HISTORY.md
@@ -4,6 +4,7 @@ History
 ### 0.5.2 (unreleased)
 
 * Support the `dangerouslySetInnerHTML` attribute.
+* Added `StateT` extension `liftR` to `ReactST` and deprecated `runState` methods handling `StateT` directly.
 
 ### 0.5.1 ([commit log](https://github.com/japgolly/scalajs-react/compare/v0.5.0...v0.5.1))